Fixed warnings and moved filter implementation to filters for AB#14587
